Added figure 4 to makefile and refactored scaling analysis function  - Added figure 4 to makefile  - Moved figure4.py analysis functions to dataImport and predict
